% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_two_dfs_with_new_column.R
\name{bind_two_dfs_with_new_column}
\alias{bind_two_dfs_with_new_column}
\title{Bind two dataframes with a new column}
\usage{
bind_two_dfs_with_new_column(df1, df2, new_column, value1, value2)
}
\arguments{
\item{df1}{Data.frame 1}

\item{df2}{Data.frame 2}

\item{new_column}{A new column name.}

\item{value1}{A value that indicates data.frame1 in the new column.}

\item{value2}{A value that indicates data.frame2 in the new column.}
}
\value{
A combined dataframe which has a new column.
}
\description{
Bind two dataframes with a new column
}
