% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_logit.R
\name{interpret_logit}
\alias{interpret_logit}
\title{Recalculate logistic model estimates in percentage terms to help interpretation}
\usage{
interpret_logit(model)
}
\arguments{
\item{model}{An outcome of a logistic regression model.}
}
\value{
A new column named "likelihood" is added to the the logistic regression model outcome.
}
\description{
Recalculate logistic model estimates in percentage terms to help interpretation
}
