% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_here.R
\name{make_here}
\alias{make_here}
\title{make_here saves typing time to parse a relative or absolute file path into here function arguments}
\usage{
make_here(filepath)
}
\arguments{
\item{filepath}{A relative or absolute file path}
}
\value{
A specified here function
}
\description{
make_here saves typing time to parse a relative or absolute file path into here function arguments
}
